ms_mailer/requirements.txt
connexion==2.6.0
Flask==1.1.1
flask-cors==3.0.8
Flask-Injector==0.12.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
flask-serialize==1.1.9
Flask-SQLAlchemy==2.4.1
swagger-ui-bundle==0.0.6
psycopg2-binary==2.8.4